requirements.txt
fastapi
peewee
uvicorn
python-jose
pytest
python-multipart
python-dotenv